Moved helper files to utils. Normalized naming through src root and utils. Moved tests
