+++ src/BetaWarning.js
