+++ src/containers/HomePage.js
@@ -13,7 +13,7 @@ import {
 import { Link } from 'react-router-dom';
 
 import { MainMenu } from '../MainMenu';
-import FixedMenu from '../fixedMenu';
+import FixedMenu from '../FixedMenu';
 import logo_stacked from '../images/logo_stacked.svg';
 
 class HomePage extends Component {
