--- src/helpers/validators.js
@@ -1,15 +1,0 @@
-
-var getMissingProps = function ( client, requiredProps ) {
-
-	var missingProps = [];
-
-	for ( let propi = 0; propi < requiredProps.length; propi++ ) {
-		let key = requiredProps[ propi ];
-		if ( client[ key ] === undefined ) { missingProps.push( key ); }
-	}
-
-	return missingProps;
-};  // End getMissingProps
-
-
-export { getMissingProps };
