+++ src/programs/state/massachusetts/template.js
@@ -17,10 +17,10 @@
 // import { SOME_DATA } from '../../../data/state/massachusetts/2018/SOME_DATA.js';
 
 // UTILITIES
-// import { roundMoney } from '../../../helpers/math';
+// import { roundMoney } from '../../../utils/math';
 
 // TO RETURN
-import { Result } from '../../../helpers/Result';
+import { Result } from '../../../utils/Result';
 
 
 /**
