+++ src/test/App.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import App from './App';
+
+import App from '../App';
 
 it('renders without crashing', () => {
   shallow(<App />);
